Fix Issue #89
This change fixes Issue #89 by making sure the player has followers. If the player has no followers, there's no point in opening the list, and if they do have followers, but it is equal to zero, then the player has not yet loaded, and it'll cause a NullPointerException if the list gets opened, which this prevents.