Fix Dribbble shot layout issue when comment field visible.
Whilst viewing shots with short descriptions or no comments, the
comment entry field is visible in the initial layout. As this field
is focusable, it requests the keyboard to show. This seems to interfere
with the entery transition, leaving it in a bad state.

This work-around places a dummy, focasable View before the RecyclerView
(which holds the comment entry field) to prevent the issue.

Fixes #142