Fix ElasticDragDismissFrameLayout bump on fast scroll
As described in issue #185 and in [this][YT] video, this
widget behave really weirdly when the scroll is performed
quickly, like a beginning of a fling gesture.

This is due to the 'returning' animation made in onStopNestedScroll.
Indeed, this method is called by the framework when ACTION_DOWN is
performed as a 'defensive cleanup for new gestures'.

This commit fixes it by retrieving the last actionEvent from
onInterceptTouchEvent, and perform the animation only if the last
actionEvent is not a ACTION_DOWN event.

Closes #185

[YT]: https://youtu.be/t-kYvsXIDoM