Ignore errors loading font in CutoutTextView.
We're using the sync method of loading the font, which apparently
fails if the font is not available (non-GMS device?). We could
try loading it async but Roboto Mono is used extensively elsewhere
via downloadable fonts, so if it's not available here, then it's
unlikely to ever be. Instead just ignore the error.

Fixes #390