Only cancel the children of the ShotsRepository.
ShotsRepository is a singleton that lives in the scope of the app. Once cancelAllSearches was called, no other calls could have been triggered, because the parent job was canceled (once a job was cancelled it cannot be active again. See https://kotlin.github.io/kotlinx.coroutines/kotlinx-coroutines-core/kotlinx.coroutines/-job/index.html ).
So for now, we're just cancelling the children.

Future, ideal fix - remove the job launching from the Repository